requirements.txt
accelerate==1.6.0
albucore==0.0.14
albumentations==1.4.14
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
comet_ml==3.49.7
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.13
decorator==5.2.1
diffusers==0.30.2
distlib==0.3.9
docker-pycreds==0.4.0
dulwich==0.22.8
easydict==1.13
eval_type_backport==0.2.2
everett==3.1.0
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
identify==2.6.9
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
insightface==0.7.3
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git@62e4af9d0c1ac7d5f8dd386a0ccf2211346af1a2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
peft==0.15.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pre_commit==4.2.0
prettytable==3.16.0
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentry-sdk==2.25.1
setproctitle==1.3.5
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.2.0
torchmetrics==1.7.1
torchvision==0.17.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triton==2.2.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
virtualenv==20.30.0
wandb==0.19.9
wcwidth==0.2.13
wrapt==1.17.2
wurlitzer==3.1.1
zipp==3.21.0
